requirements.txt
instagrapi
schedule
python-dotenv
Pillow
